IntCodeCpu supports big numbers using System.Numerics. Add relative address support.

+++ AdventOfCode_2019/Week01/02/Day02.cs
@@ -68,7 +68,7 @@ namespace AdventOfCode_2019.Week01
             logger.LogDebug($"Executed {cpu.Steps} operations.");
             logger.LogDebug($"End {cpu.DumpMemory()}");
 
-            return cpu[0];
+            return (int)cpu[0];
         }
     }
 }
